+++ packages/iobio-react-components/src/utils/statisticsClient/statisticsTools.mts
@@ -22,7 +22,7 @@
 import { DataBroker } from 'iobio-charts/data_broker.js';
 import fs from 'node:fs';
 import readline from 'node:readline/promises';
-import { calculateMeanCoverage, getBamStatistics } from '../iobioHelpers.mts';
+import { calculateMeanCoverage, getBamStatistics } from '../iobioTools.mts';
 import { type StatsOutput } from '../iobioTypes.ts';
 
 export type CompleteCallback = (stats: StatsOutput) => Promise<void>;
